requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.0
pillow>=9.0.0
numpy>=1.24.0
scipy>=1.10.0
opencv-python>=4.8.0
scikit-image>=0.20.0
matplotlib>=3.6.0
seaborn>=0.12.0
tqdm>=4.64.0
einops>=0.6.0
omegaconf>=2.3.0
hydra-core>=1.3.0
wandb>=0.15.0
tensorboard>=2.13.0
gradio>=3.40.0
streamlit>=1.25.0
PyPDF2>=3.0.1
